extract controller into separated configuration
