+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/Configuration.kt
@@ -1,6 +1,7 @@
 package github.com.hukuta94.delivery.api.startup.configuration
 
 import github.com.hukuta94.delivery.api.startup.configuration.adapter.KafkaConfiguration
+import github.com.hukuta94.delivery.api.startup.configuration.adapter.http.ControllerConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.adapter.orm.OrmRepositoryConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.courier.CourierConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.order.OrderConfiguration
@@ -8,10 +9,8 @@ import github.com.hukuta94.delivery.api.startup.configuration.scheduler.Schedule
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
-import org.springframework.web.servlet.config.annotation.EnableWebMvc
 
 @Configuration
-@EnableWebMvc
 @EnableAutoConfiguration
 @Import(
     OrderConfiguration::class,
@@ -22,5 +21,6 @@ import org.springframework.web.servlet.config.annotation.EnableWebMvc
     DomainEventHandlerConfiguration::class,
     KafkaConfiguration::class,
     OrmRepositoryConfiguration::class,
+    ControllerConfiguration::class,
 )
 open class Configuration
