+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/http/ControllerConfiguration.kt
@@ -1,0 +1,13 @@
+package github.com.hukuta94.delivery.api.startup.configuration.adapter.http
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
+import org.springframework.web.servlet.config.annotation.EnableWebMvc
+
+@EnableWebMvc
+@Configuration
+@Import(
+    OrderControllerConfiguration::class,
+    CourierControllerConfiguration::class,
+)
+open class ControllerConfiguration
