+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/courier/CourierConfiguration.kt
@@ -6,6 +6,5 @@ import org.springframework.context.annotation.Import
 @Configuration
 @Import(
     CourierUseCaseConfiguration::class,
-    CourierControllerConfiguration::class,
 )
 open class CourierConfiguration
