+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/order/OrderConfiguration.kt
@@ -6,6 +6,5 @@ import org.springframework.context.annotation.Import
 @Configuration
 @Import(
     OrderUseCaseConfiguration::class,
-    OrderControllerConfiguration::class,
 )
 open class OrderConfiguration
